+++ ChessEngine/src/main/java/org/example/Engine/BoardRepresentation/BoardFormats/BitBoardsRepresentation.java
@@ -7,13 +7,13 @@ public class BitBoardsRepresentation implements Format {
     public long[] bitBoards = new long[15];
 
     @Override
-    public void addPieceOnSquare(short square, short color, short piece) {
+    public void addPieceOnSquare(byte square, byte color, byte piece) {
         bitBoards[piece | color] |= (1L << square-1);
         bitBoards[color] |= (1L << square-1);
     }
 
     @Override
-    public void deletePieceOnSquare(short square, short color, short piece) {
+    public void deletePieceOnSquare(byte square, byte color, byte piece) {
         bitBoards[piece | color] &= ~(1L << square-1);
         bitBoards[color] &= ~(1L << square-1);
     }
