+++ ChessEngine/src/main/java/org/example/Engine/BoardRepresentation/BoardFormats/CountRepresentation.java
@@ -4,20 +4,20 @@ import java.util.Arrays;
 
 public class CountRepresentation implements Format {
 
-    public short[] pieces = new short[16];
+    public byte[] pieces = new byte[16];
 
     @Override
-    public void addPieceOnSquare(short square, short color, short piece) {
+    public void addPieceOnSquare(byte square, byte color, byte piece) {
         pieces[color | piece]++;
     }
 
     @Override
-    public void deletePieceOnSquare(short square, short color, short piece) {
+    public void deletePieceOnSquare(byte square, byte color, byte piece) {
         pieces[color | piece]--;
     }
 
     @Override
     public void clearBoard() {
-        Arrays.fill(pieces, (short) 0);
+        Arrays.fill(pieces, (byte) 0);
     }
 }
