+++ ChessEngine/src/main/java/org/example/Engine/BoardRepresentation/BoardFormats/Format.java
@@ -1,7 +1,7 @@
 package org.example.Engine.BoardRepresentation.BoardFormats;
 
 public interface Format {
-    void addPieceOnSquare(short square, short color, short piece);
-    void deletePieceOnSquare(short square, short color, short piece);
+    void addPieceOnSquare(byte square, byte color, byte piece);
+    void deletePieceOnSquare(byte square, byte color, byte piece);
     void clearBoard();
 }
