+++ ChessEngine/src/test/java/org/example/Engine/BoardRepresentation/BoardFormats/BitBoardsRepresentationTest.java
@@ -8,10 +8,10 @@ public class BitBoardsRepresentationTest extends TestCase {
 
         public void testAddPieceOnSquare() {
             BitBoardsRepresentation bitBoards = new BitBoardsRepresentation();
-            bitBoards.addPieceOnSquare((short) 64, BLACK, PAWN);
-            bitBoards.addPieceOnSquare((short) 57, BLACK, PAWN);
-            bitBoards.addPieceOnSquare((short) 21, BLACK, PAWN);
-            bitBoards.addPieceOnSquare((short) 1, BLACK, PAWN);
+            bitBoards.addPieceOnSquare((byte) 64, BLACK, PAWN);
+            bitBoards.addPieceOnSquare((byte) 57, BLACK, PAWN);
+            bitBoards.addPieceOnSquare((byte) 21, BLACK, PAWN);
+            bitBoards.addPieceOnSquare((byte) 1, BLACK, PAWN);
 
             displayBitBoard(bitBoards.bitBoards[BLACK | PAWN]);
             assertEquals(-9151314442815799295L, bitBoards.bitBoards[BLACK | PAWN]);
